pyproject.toml
[project]
name = "daily-digest-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "fastapi>=0.113.0",
    "ipykernel>=7.1.0",
    "openai>=2.15.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.31",
    "streamlit>=1.54.0",
    "uvicorn>=0.30.3",
]

[dependency-groups]
dev = [
    "notebook>=7.5.2",
]
